earthquake-dashboard/requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
requests
paho-mqtt
python-dotenv
twilio
geocoder
pyserial
